edits

+++ my_api/initialize_api.py
@@ -3,13 +3,14 @@ from fastapi import APIRouter
 from rag_core.get_answer_from_model import get_answer_from_model
 from pydantic import BaseModel
 from typing import List
+import os
 router = APIRouter(prefix="/api")
 
 class QueryRequest(BaseModel):
     chunks: int
     numofresults: int
     question: str
-    filepaths: List[str]
+    filenames: List[str]
 @router.get("/")
 def root():
     return {"message": "Hello tt"}
@@ -22,7 +23,9 @@ async def postupload(
     for file in files:
         print(f"Received file: {file.filename}, size: {file.size} bytes")
         file_location = f"rag_core/src/{file.filename}"
-
+        if os.path.exists(file_location):
+            print(f"File {file.filename} already exists at {file_location}. Skipping save.")
+            continue 
         try:
             # Save the file
             with open(file_location, "wb") as f:
@@ -43,7 +46,8 @@ async def postq(query_request: QueryRequest):
     chunks = query_request.chunks
     numofresults = query_request.numofresults
     question = query_request.question
-    filepaths = [f"rag_core/src/{file}" for file in query_request.filepaths]
+    filepaths = [f"rag_core/src/{file}" for file in query_request.filenames]
 
     result = get_answer_from_model(filepaths, chunks, numofresults, question)
+    print(result)
     return result
